[puppi] added functionality to use PU-proxy different from N_PV

+++ CommonTools/PileupAlgos/interface/PuppiContainer.h
@@ -10,7 +10,7 @@ public:
   PuppiContainer(const edm::ParameterSet &iConfig);
   ~PuppiContainer();
   void initialize(const std::vector<RecoObj> &iRecoObjects);
-  void setNPV(int iNPV) { fNPV = iNPV; }
+  void setNPV(double iNPV) { fNPV = iNPV; }
 
   std::vector<PuppiCandidate> const &pfParticles() const { return fPFParticles; }
   std::vector<PuppiCandidate> const &pvParticles() const { return fChargedPV; }
@@ -61,7 +61,7 @@ protected:
   double fPtMaxNeutrals;
   double fPtMaxNeutralsStartSlope;
   int fNAlgos;
-  int fNPV;
+  double fNPV;
   double fPVFrac;
   std::vector<PuppiAlgo> fPuppiAlgo;
 };
